requirements.txt
amqp==2.5.2
ansible==2.7.12
asn1crypto==0.24.0
backcall==0.1.0
bcrypt==3.1.4
billiard==3.6.2.0
celery==4.4.0
celery-with-redis==3.0
certifi==2019.11.28
cffi==1.12.2
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==2.6.1
decorator==4.4.0
Django==2.2.1
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-cors-headers==3.0.1
django-cors-middleware==1.3.1
django-filter==2.2.0
django-filters==0.2.1
django-redis==4.10.0
django-timezone-field==4.1.2
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
drf-extensions==0.6.0
idna==2.8
importlib-metadata==1.5.0
IPy==1.0
ipython==7.4.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10
kombu==4.6.7
Markdown==3.1
MarkupSafe==1.1.1
multi-key-dict==2.0.3
mysqlclient==1.4.2.post1
paramiko==2.6.0
parso==0.3.4
pbr==5.4.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyasn1==0.4.5
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
PyMySQL==0.9.3
PyNaCl==1.3.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-jenkins==1.4.0
pytz==2018.9
PyYAML==5.1
raven==6.7.0
redis==3.4.1
requests==2.21.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
simplegeneric==0.8.1
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.2
typing-extensions==3.10.0.0
uritemplate==3.0.0
urllib3==1.24.1
uWSGI==2.0.18
vine==1.3.0
wcwidth==0.1.7
zipp==2.2.0
